requirements.txt
fastapi
pyswisseph
sqlalchemy
psycopg2-binary
python-dotenv
uvicorn
pydantic
swisseph
passlib[bcrypt]
pyjwt
